# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type EmployeeStatus = typeof EmployeeStatus[keyof typeof EmployeeStatus];


export const EmployeeStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Employee {
  id: string;
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  status: EmployeeStatus;
  createdAt: string;
  updatedAt: string;
}

export type EmployeeInputStatus = typeof EmployeeInputStatus[keyof typeof EmployeeInputStatus];


export const EmployeeInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface EmployeeInput {
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  status?: EmployeeInputStatus;
}

export type EmployeeUpdateStatus = typeof EmployeeUpdateStatus[keyof typeof EmployeeUpdateStatus];


export const EmployeeUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface EmployeeUpdate {
  name?: string;
  department?: string;
  position?: string;
  status?: EmployeeUpdateStatus;
}

export type AttendanceLogType = typeof AttendanceLogType[keyof typeof AttendanceLogType];


export const AttendanceLogType = {
  checkin: 'checkin',
  checkout: 'checkout',
} as const;

export interface AttendanceLog {
  id: string;
  employeeId: string;
  employeeCode: string;
  employeeName: string;
  department: string;
  type: AttendanceLogType;
  timestamp: string;
}

export interface AttendanceAction {
  employeeCode: string;
}

export interface AttendanceLogsResponse {
  logs: AttendanceLog[];
  total: number;
  offset: number;
  limit: number;
}

export interface ActiveEmployee {
  employeeId: string;
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  checkedInAt: string;
}

export interface DashboardStats {
  totalEmployees: number;
  activeNow: number;
  checkInsToday: number;
  checkOutsToday: number;
  presentToday: number;
}

export type ListEmployeesParams = {
search?: string;
department?: string;
status?: ListEmployeesStatus;
};

export type ListEmployeesStatus = typeof ListEmployeesStatus[keyof typeof ListEmployeesStatus];


export const ListEmployeesStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListAttendanceLogsParams = {
employeeId?: string;
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
type?: ListAttendanceLogsType;
limit?: number;
offset?: number;
};

export type ListAttendanceLogsType = typeof ListAttendanceLogsType[keyof typeof ListAttendanceLogsType];


export const ListAttendanceLogsType = {
  checkin: 'checkin',
  checkout: 'checkout',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};