# lib/api-zod/src/generated/types/activeEmployee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveEmployee {
  employeeId: string;
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  checkedInAt: Date;
}