# lib/api-zod/src/generated/types/attendanceLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceLogType } from './attendanceLogType';

export interface AttendanceLog {
  id: string;
  employeeId: string;
  employeeCode: string;
  employeeName: string;
  department: string;
  type: AttendanceLogType;
  timestamp: Date;
}