# lib/api-zod/src/generated/types/attendanceLogsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceLog } from './attendanceLog';

export interface AttendanceLogsResponse {
  logs: AttendanceLog[];
  total: number;
  offset: number;
  limit: number;
}