# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalEmployees: number;
  activeNow: number;
  checkInsToday: number;
  checkOutsToday: number;
  presentToday: number;
}