# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: string;
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  status: EmployeeStatus;
  createdAt: Date;
  updatedAt: Date;
}