# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputStatus } from './employeeInputStatus';

export interface EmployeeInput {
  employeeCode: string;
  name: string;
  department: string;
  position: string;
  status?: EmployeeInputStatus;
}