# lib/api-zod/src/generated/types/employeeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Attendance control SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeUpdateStatus } from './employeeUpdateStatus';

export interface EmployeeUpdate {
  name?: string;
  department?: string;
  position?: string;
  status?: EmployeeUpdateStatus;
}